models/menu_translator/requirements.txt
# Menu Translator - ML Models
# Python 3.10+

# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0

# Translation - HuggingFace Transformers
transformers>=4.40.0
sentencepiece>=0.2.0
sacremoses>=0.1.0

# OCR
paddlepaddle>=2.6.0
paddleocr>=2.8.0

# NLP
langdetect>=1.0.9

# Image processing
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.9.0

# Metrics
prometheus-client>=0.20.0

# Utilities
tqdm>=4.65.0
